requirements.txt

streamlit
openai
mysql-connector-python
pandas
